LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.measure.emHeights-expected.txt

2d.text.measure.emHeights
Testing emHeights
Actual output:

FAIL Testing emHeights assert_equals: ctx.measureText('A').emHeightAscent === 30 (got 69[number], expected 30[number]) expected 30 but got 69
